requirements.txt
# Python Dependencies for Automotive Predictive Maintenance
# Phase 1 + Phase 2 (ClickHouse Integration & FastAPI)
# Python 3.11+ required

# ============================================================================
# PHASE 1 - MQTT → Kafka → Data Cleaning
# ============================================================================

# MQTT Client
paho-mqtt==1.6.1

# Kafka Client (Async)
aiokafka==0.12.0

# Data Validation
pydantic==2.5.0

# Environment Variables
python-dotenv==1.0.0

# Structured Logging
loguru==0.7.2

# ============================================================================
# PHASE 2 - ClickHouse Integration & FastAPI
# ============================================================================

# ClickHouse Database Driver
clickhouse-driver==0.2.6

# FastAPI Web Framework
fastapi==0.104.1

# ASGI Server for FastAPI
uvicorn[standard]==0.24.0

# HTTP Client (for testing API)
httpx==0.25.2

# Additional Pydantic Features
pydantic-settings==2.1.0

# ============================================================================
# PHASE 3 - ML Predictive Maintenance
# ============================================================================

# Core ML Libraries
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# Gradient Boosting (optional, for advanced models)
xgboost==2.0.3

# Model Serialization
joblib==1.3.2

# Feature Engineering
scipy==1.11.4

# ============================================================================
# PHASE 4 - Multi-Agent Ecosystem
# ============================================================================

# MongoDB Driver (Async)
motor==3.3.2

# Anomaly Detection for UEBA
pyod==1.1.3

# Optional: Notification Services
twilio==8.10.0  # For SMS/WhatsApp (optional)
sendgrid==6.11.0  # For Email (optional)

# ============================================================================
# PHASE 5 - Manufacturing Feedback & RCA with ClickHouse Analytics
# ============================================================================

# Already included above but ensuring compatibility:
# clickhouse-driver==0.2.6  (for historical analytics)
# pandas==2.1.3  (for trend analysis)
# numpy==1.26.2  (for statistical calculations)
# motor==3.3.2  (for MongoDB integration)
# aiokafka==0.12.0  (for Kafka streaming)

# ============================================================================
# PHASE 6 - UEBA (User and Entity Behavior Analytics) with ElasticSearch
# ============================================================================

# ElasticSearch Client (Async)
elasticsearch==8.11.0
elastic-transport==8.11.0

# Already included above but critical for Phase 6:
# pyod==1.1.3  (Isolation Forest + LOF for anomaly detection)
# pandas==2.1.3  (for feature engineering)
# numpy==1.26.2  (for statistical calculations)
# scikit-learn==1.3.2  (for StandardScaler)
# motor==3.3.2  (for MongoDB)
# aiokafka==0.12.0  (for Kafka)
# fastapi==0.104.1  (for UEBA Dashboard API)
# uvicorn[standard]==0.24.0  (for running API)

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Formatting
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Monitoring (for future phases)
prometheus-client==0.19.0

# ============================================================================
# NLP ENHANCEMENTS (Optional - Phase 7)
# ============================================================================

# Core NLP Libraries
# openai==1.3.0  # Uncomment when using OpenAI API
# langchain==0.1.0  # Uncomment when implementing NLP features
# langchain-openai==0.0.2
transformers==4.36.0
torch==2.1.0

# NLP Utilities
spacy==3.7.0
nltk==3.8.1
vaderSentiment==3.3.2
textblob==0.17.1

# Speech Processing (Optional)
# openai-whisper==20231117  # Large model, uncomment if needed
# gTTS==2.4.0
# SpeechRecognition==3.10.0

# Translation (Optional)
# googletrans==4.0.0rc1
# langdetect==1.0.9

# Vector Databases (Optional - for knowledge bases)
# pinecone-client==2.2.4
# chromadb==0.4.18

# Explainable AI (Optional)
# shap==0.44.0
# lime==0.2.0.1

# Document Processing (Optional)
# python-docx==1.1.0
# pdfplumber==0.10.3